+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added the option to change the log level or suppress logging of unhandled errors which are, in fact,
+  handled by Rails (e.g. 404 Not Found).
 
 ### Changed
 - The initial context is now fully dynamic; it can be either a Hash or a lambda/Proc returning a Hash.